Fix the sign-out redirect loop and a react-doctor false positive
RequireAuth's <Navigate> compared its whole props object by reference
on every render to decide whether to re-navigate; JSX always builds a
fresh object, so it fired on every render and fed off its own
useLocation() read, looping. Replaced it with an effect gated on a
stable boolean that reads the path imperatively.

Also split dashboardView out of dashboard.tsx into its own module:
react-doctor's only-export-components rule flagged it as new on the
PR diff even though the line was unchanged, since the file itself had
other edits.
